requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.1
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Faker==11.3.0
future==0.18.2
html5lib==1.1
idna==3.3
Pillow==9.0.1
PyJWT==2.3.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
reportlab==3.6.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.8
webencodings==0.5.1
xhtml2pdf==0.2.5
